refactor the structure, separate api from web endpoints
