+++ app/common/auth.py
@@ -1,0 +1,16 @@
+from flask import g
+from flask_httpauth import HTTPBasicAuth
+from app.models.user import User
+
+
+auth = HTTPBasicAuth()
+@auth.verify_password
+def verify_password(username_or_token, password):
+    user = User.verify_auth_token(username_or_token)
+    if not user:
+        # try to authenticate with username/password
+        user = User.query.filter_by(username=username_or_token).first()
+        if not user or not user.check_password(password):
+            return False
+    g.user = user
+    return True
