+++ configuration/structure.txt
@@ -1,0 +1,29 @@
+iot\
+    app\
+        api\
+            resources\
+                __init__.py
+                token.py
+                user.py
+                device.py
+            __init__.py
+            api.py
+        common\
+            __init__.py
+            auth.py
+            db.py
+        models\
+            __init__.py
+            user.py
+            device.py
+        static\
+        frontend\
+            __init__.py
+            controllers.py
+        __init__.py
+        app.py
+    migration\
+    configurations\
+    iot.py
+    mqtt.py
+    config.py
